// src/theme/palette.types.ts
import type {ColorValue} from 'react-native';

export interface Palette {
  [key: string]: ColorValue;
  primary0: ColorValue;
  primary10: ColorValue;
  primary20: ColorValue;
  primary30: ColorValue;
  primary40: ColorValue;
  primary60: ColorValue;
  primary80: ColorValue;
  primary90: ColorValue;
  primary100: ColorValue;
  secondary0: ColorValue;
  secondary10: ColorValue;
  secondary20: ColorValue;
  secondary30: ColorValue;
  secondary40: ColorValue;
  secondary60: ColorValue;
  secondary80: ColorValue;
  secondary90: ColorValue;
  secondary100: ColorValue;
  tertiary0: ColorValue;
  tertiary10: ColorValue;
  tertiary20: ColorValue;
  tertiary30: ColorValue;
  tertiary40: ColorValue;
  tertiary60: ColorValue;
  tertiary80: ColorValue;
  tertiary90: ColorValue;
  tertiary100: ColorValue;
  error10: ColorValue;
  error20: ColorValue;
  error30: ColorValue;
  error40: ColorValue;
  error80: ColorValue;
  error90: ColorValue;
  error100: ColorValue;
  neutral0: ColorValue;
  neutral4: ColorValue;
  neutral6: ColorValue;
  neutral10: ColorValue;
  neutral12: ColorValue;
  neutral17: ColorValue;
  neutral20: ColorValue;
  neutral22: ColorValue;
  neutral24: ColorValue;
  neutral30: ColorValue;
  neutral40: ColorValue;
  neutral60: ColorValue;
  neutral87: ColorValue;
  neutral90: ColorValue;
  neutral92: ColorValue;
  neutral94: ColorValue;
  neutral95: ColorValue;
  neutral96: ColorValue;
  neutral98: ColorValue;
  neutral100: ColorValue;
  neutralVariant30: ColorValue;
  neutralVariant50: ColorValue;
  neutralVariant60: ColorValue;
  neutralVariant80: ColorValue;
}